Split Gauges into subtypes.
SetGauge is a kind of gauge that can be set() to a given value.

IncDecGauge is a kind of gauge that can be incremented or
decremented. You can't set() an IncDecGauge. This guarantees that an
inc followed by a dec (or vice-versa) always ends up with an unchanged
counter. I will add a RAII-based util to inc() and dec() them when in
the scope of a function in a later commit.

CallbackGauges will come soon, too.