Run ASAN without -DADDRESS_SANITIZER.
This macro is used by protbuf's arenas to instrument the code in
certain ways, but this requires headers that are only available under
clang. Since Bazel has no clang toolchain right now by default, we use
gcc, and we need to disable this. All tests still pass under Asan (but
I don't use protobufs yet so we'll see how this goes in the future).