Protect all values in a histogram with a single mutex.
Previously we relied on multiple atomic counters. This doesn't work as
there is an invariant that a value observed in the first bucket is
always observed in the next bucket. We could solve that by iterating
in reverse order over the buckets, but a histogram also maintains a
total of all observed values and that should be in sync with the +Inf
bucket's count (so the average is not thrown off). Protecting all
values with a single mutex solves this.

There is a possible optimization, which is to duplicate the total
count of observed values outside of the +inf bucket, and protect that
count and the total sum of observer values with a mutex, and let each
bucket be its own atomic<double>, iterating in reverse order to
increment them.