Use locally scoped regex objects to test label/metric names.
Using global objects is risky as metrics are created as globals and
this creates a global initialization order race. Using local regex
objects is more expensive but since this only applies to metric
creation, this is acceptable (it may slightly slow down start times
but the runtime cost is zero).